backend/pyproject.toml
[project]
name = "fair-way"
version = "0.8.0"
description = "\"LLM Based Tool for FAIR Data Assessment\""
authors = [
    {name = "Anmol Sharma",email = "anmol.sharma@rwth-aachen.de"}
]
readme = "README.md"
requires-python = ">=3.12 <3.13"
dependencies = [
    "celery (==5.4.0)",
    "fastapi (>=0.115.7,<0.116.0)",
    "ollama (>=0.4.7,<0.5.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "langchain-text-splitters (>=0.3.5,<0.4.0)",
    "redis (>=5.2.1,<6.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "uvicorn (>=0.34.0,<0.35.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"

[tool.poetry]
package-mode = false